refactor: change code structure for controllers and change email middleware abstraction, add updateUser route
