+++ src/controllers/User/registerUser.controller.ts
@@ -5,28 +5,11 @@ import { createUserService } from "../../services/User/registerUser.service";
 import { AppError } from "../../utils/errorHandler.util";
 
 export const registerUserController = async (req: Request, resp: Response) => {
-    try {
-        const userData = req.body;
-        const addressData = req.body.address;
-    
-        const parseUserData = createdUserSchema.parse(userData);
-    
-        const createdNewUser = await createUserService(parseUserData,addressData );
-        console.log(createdNewUser,"createdUser");
-        return resp.status(statusSuccess.CREATED).json(createdNewUser);
-    } catch (error) {
-        console.log(error);
-        throw new AppError(`internal server error ${error} `,statusError.SERVER_ERROR);
-    }
-   
+    const userData = req.body;
+    const addressData = req.body.address;
+
+    const parseUserData = createdUserSchema.parse(userData);
+
+    const createdNewUser = await createUserService(parseUserData, addressData);
+    return resp.status(statusSuccess.CREATED).json(createdNewUser);
 };
-
-// export const registerUserController = async (req: Request, res: Response)=> {
-    
-//     const userData = req.body;
-//     const address = req.body.address;
-
-//     const createdUser = await createUserService(userData, address);
-
-//     return res.status(statusSuccess.CREATED).json(createdUser);
-// };
