+++ src/routes/User/user.routes.ts
@@ -1,35 +1,28 @@
 import { Router } from "express";
-import { createUserAdressController } from "../../controllers/User/Address/createUserAddress.controller";
 import { deleteUserByIdController } from "../../controllers/User/deleteUserbyId.controller";
 import { getUserByIdController } from "../../controllers/User/getUserByid.controller";
 import { registerUserController } from "../../controllers/User/registerUser.controller";
 import { userLoginController } from "../../controllers/User/userLogin.controller";
-import { verifyTokenIsValid } from "../../middlewares/Global/checktoken.middleware";
-import { validateDataMiddleware } from "../../middlewares/Global/validatedData";
-import { checkEmailMiddleware } from "../../middlewares/User/checkEmail.middleware";
-import { createdAdress } from "../../schemas/User/userAddress.schema";
+import { verifyTokenMiddleware } from "../../middlewares/Global/verifyToken.middleware";
+import { validateDataMiddleware } from "../../middlewares/Global/validateData.middleware";
 import { createdLoginSchema } from "../../schemas/User/userLogin.schema";
+import { checkEmailMiddleware } from "../../middlewares/User/checkEmail.middleware";
+import { updateUserController } from "../../controllers/User/updateUser.controller";
 
 const userRouter: Router = Router();
 
-userRouter.post(
-    "",
-   
-    checkEmailMiddleware,
-    registerUserController
-);
-userRouter.get("", verifyTokenIsValid, getUserByIdController);
+userRouter.post("", checkEmailMiddleware, registerUserController);
+
 userRouter.post(
     "/login",
     validateDataMiddleware(createdLoginSchema),
     userLoginController
 );
-userRouter.post(
-    "/address",
-    verifyTokenIsValid,
-    validateDataMiddleware(createdAdress),
-    createUserAdressController
-);
-userRouter.delete("", verifyTokenIsValid, deleteUserByIdController);
+
+userRouter.get("", verifyTokenMiddleware, getUserByIdController);
+
+userRouter.patch("", verifyTokenMiddleware, updateUserController);
+
+userRouter.delete("", verifyTokenMiddleware, deleteUserByIdController);
 
 export default userRouter;
